Extract remote workspace browser harness
